bot/requirements.txt
discord.py
mysql-connector-python
requests
Pillow
pytz